Shapes
The Way has no true shape,
And therefore none can control it.
If a ruler could control the Way
All things would follow
In harmony with his desire,
And sweet rain would fall,
Effortlessly slaking every thirst.
The Way is shaped by use,
But then the shape is lost.
Do not hold fast to shapes
But let sensation flow into the world
As a river courses down to the sea.
		-- Lao Tse, "Tao Te Ching"